test(Effect): move Effectable regression coverage
